logs has been removed.

+++ EMailService/Modal/Payroll/PayrollMonthlyDetail.cs
@@ -20,6 +20,6 @@ namespace EMailService.Modal.Payroll
         public DateTime ExecutedOn { set; get; }
         public int CompanyId { set; get; }
         public int TotalEmployees { set; get; }
-        public int LOP { get; set; }
+        public decimal LOP { get; set; }
     }
 }
